requirements.txt
streamlit
torch
huggingface_hub
pillow
numpy